move to _bloch
