+++ qiskit/tools/visualization/_bloch.py
